clips: address Review Agent feedback on the blob route
Three real issues flagged by builder-io-integration on the previous
commits. All addressed by moving the route and hardening it:

1. Password protection was bypassed — the blob route checked visibility
   and expiry but never validated `recordings.password`, while
   public-recording.get.ts does. A viewer who knew the video URL could
   skip the password gate.

2. Anonymous share playback was broken — the route lived under
   /api/uploads/* which the auth plugin guards, so anonymous viewers on
   /share/:id couldn't fetch video bytes when the recording was public.

3. HTTP Range parsing was not RFC 9110 compliant — suffix ranges
   (bytes=-N, "last N bytes") were parsed as bytes=0-N and served the
   wrong segment. Oversized `end` values returned 416 instead of
   clamping to total-1.

Fixes:

- Move the route to /api/video/:recordingId and add `/api/video` to
  publicPaths in server/plugins/auth.ts. /api/uploads/* stays protected
  so the chunk-upload POST endpoints remain auth-gated; the read-only
  blob serve happens in its own namespace.
- Enforce the password gate the same way public-recording.get.ts does:
  owners bypass (they set it); everyone else must pass ?password=<pw>
  as a query parameter or get a 401 with { passwordRequired: true }.
- Handle suffix ranges (bytes=-N → [total-N, total-1]) correctly and
  clamp oversized `end` values to total-1 rather than 416'ing.
- Update finalize-recording.ts to set videoUrl=/api/video/:id.
- Update request-transcript.ts's dev-fallback detection to accept
  both the new /api/video/:id path and the legacy /api/uploads/:id/blob
  shape for any old rows still around.
